+++ ew/farm.py
@@ -10,7 +10,6 @@ from . import rolemgr as ewrolemgr
 
 from .user import EwUser
 from .market import EwMarket
-from .food import EwFood
 from .item import EwItem
 from .slimeoid import EwSlimeoid
 from .district import EwDistrict
@@ -101,34 +100,6 @@ class EwFarm:
 				self.sow_life_state,
 			)
 		)
-
-class EwFarmAction:
-	id_action = 0
-
-	action = ""
-	
-	str_check = ""
-
-	str_execute = ""
-
-	str_execute_fail = ""
-
-	aliases = []
-
-	def __init__(self,
-		id_action = 0,
-		action = "",
-		str_check = "",
-		str_execute = "",
-		str_execute_fail = "",
-		aliases = []
-	):
-		self.id_action = id_action
-		self.action = action
-		self.str_check = str_check
-		self.str_execute = str_execute
-		self.str_execute_fail = str_execute_fail
-		self.aliases = aliases
 
 
 """
