+++ ew/move.py
@@ -8,11 +8,12 @@ import discord
 from copy import deepcopy
 
 from . import utils as ewutils
-from . import cmd as ewcmd
 from . import rolemgr as ewrolemgr
 from .static import cfg as ewcfg
 from . import apt as ewapt
 from . import ads as ewads
+from . import hunting as ewhunting
+
 
 from .user import EwUser
 from .district import EwDistrict
@@ -25,7 +26,7 @@ from .ads import EwAd
 from .item import EwItem
 from .dungeons import EwGamestate
 
-from .hunting import EwEnemy, spawn_enemy
+#from .hunting import EwEnemy
 from .worldevent import get_void_connection_pois
 
 move_counter = 0
@@ -68,240 +69,6 @@ def poi_id_to_display_name(poi_name = None):
         return poi.str_name
 
     return "the city"
-
-"""
-    Point of Interest (POI) data model
-"""
-class EwPoi:
-    # The typable single-word ID of this location.
-    id_poi = ""
-
-    # Acceptable alternative typable single-word names for this place.
-    alias = []
-
-    # The nice name for this place.
-    str_name = ""
-
-    # You find yourself $str_in $str_name
-    str_in = "in"
-
-    # You $str_enter $str_name
-    str_enter = "enter"
-
-    # A description provided when !look-ing here.
-    str_desc = ""
-
-    # (X, Y) location on the map (left, top) zero-based origin.
-    coord = None
-    coord_alias = []
-
-    # Channel name associated with this POI
-    channel = ""
-
-    # Discord role associated with this zone (control channel visibility).
-    role = None
-
-    # Role that controls LAN voice/text channel visibility for any street/subzone/district - 6/6/20
-    major_role = None
-
-    # Role that controls subzone visibility for streets/districts - 6/6/20
-    minor_role = None
-
-    # Discord permissions associated with this zone (control channel visibility) - 5/28/20
-    permissions = None
-
-    # Zone allows PvP combat and interactions.
-    pvp = True
-
-    # Factions allowed in this zone.
-    factions = []
-
-    # Life states allowed in this zone.
-    life_states = []
-
-    # If true, the zone is inaccessible.
-    closed = False
-
-    # Message shown before entering the zone fails when it's closed.
-    str_closed = None
-
-    # Vendor names available at this POI.
-    vendors = []
-
-    # The value of the district
-    property_class = ""
-
-    # If the zone is a district
-    is_district = False
-
-    # If the zone is a gang base (Juvie's Row included)
-    is_gangbase = False
-
-    # If true, the zone is a district that can be controlled/captured
-    is_capturable = False
-
-    # If it's a subzone
-    is_subzone = False
-
-    #If it's an apartment
-    is_apartment = False
-
-    # if this zone is a street within a district
-    is_street = False
-
-    # What District/street each subzone is in. Subzones could potentially have multiple mother districts if they are between streets/districts.
-    mother_districts = []
-
-    # What District each street is attatched to
-    father_district = ""
-
-    # If it's a mobile zone
-    is_transport = False
-
-    # which type of transport
-    transport_type = ""
-
-    # default line to follow, if it's a transport
-    default_line = ""
-
-    # default station to start at, if it's a transport
-    default_stop = ""
-
-    # If a transport line stops here
-    is_transport_stop = True
-
-    # which transport lines stop here
-    transport_lines = set()
-
-    # if this zone belongs to the outskirts
-    is_outskirts = False
-
-    # id for the zone's community chest, if it has one
-    community_chest = None
-
-    # if you can fish in the zone
-    is_pier = False
-
-    # if the pier is in fresh slime or salt slime
-    pier_type = None
-
-    # if the poi is part of the tutorial
-    is_tutorial = False
-
-    # whether to show ads here
-    has_ads = False
-
-    # if you can write zines here
-    write_manuscript = False
-
-    # maximum degradation - zone ceases functioning when this value is reached
-    max_degradation = 0
-
-    # dict EwPoi -> int, that defines travel times into adjacent pois
-    neighbors = None
-
-    # The topic associated with that poi's channel
-    topic = ""
-
-    # The wiki page associated with that poi
-    wikipage = ""
-
-    def __init__(
-        self,
-        id_poi = "unknown",
-        alias = [],
-        str_name = "Unknown",
-        str_desc = "...",
-        str_in = "in",
-        str_enter = "enter",
-        coord = None,
-        coord_alias = [],
-        channel = "",
-        role = None,
-        major_role = None,
-        minor_role = None,
-        permissions = None,
-        pvp = True,
-        factions = [],
-        life_states = [],
-        closed = False,
-        str_closed = None,
-        vendors = [],
-        property_class = "",
-        is_district = False,
-        is_gangbase = False,
-        is_capturable = False,
-        is_subzone = False,
-        is_apartment = False,
-        is_street = False,
-        mother_districts = [],
-        father_district = "",
-        is_transport = False,
-        transport_type = "",
-        default_line = "",
-        default_stop = "",
-        is_transport_stop = False,
-        transport_lines = None,
-        is_outskirts = False,
-        community_chest = None,
-        is_pier = False,
-        pier_type = None,
-        is_tutorial = False,
-        has_ads = False,
-        write_manuscript = False,
-        max_degradation = 10000,
-        neighbors = None,
-        topic = "",
-        wikipage = "",
-    ):
-        self.id_poi = id_poi
-        self.alias = alias
-        self.str_name = str_name
-        self.str_desc = str_desc
-        self.str_in = str_in
-        self.str_enter = str_enter
-        self.coord = coord
-        self.coord_alias = coord_alias
-        self.channel = channel
-        self.role = role
-        self.major_role = major_role
-        self.minor_role = minor_role
-        self.permissions = permissions
-        self.pvp = pvp
-        self.factions = factions
-        self.life_states = life_states
-        self.closed = closed
-        self.str_closed = str_closed
-        self.vendors = vendors
-        self.property_class = property_class
-        self.is_district = is_district
-        self.is_gangbase = is_gangbase
-        self.is_capturable = is_capturable
-        self.is_subzone = is_subzone
-        self.is_apartment = is_apartment
-        self.is_street = is_street
-        self.mother_districts = mother_districts
-        self.father_district = father_district
-        self.is_transport = is_transport
-        self.transport_type = transport_type
-        self.default_line = default_line
-        self.default_stop = default_stop
-        self.is_transport_stop = is_transport_stop
-        self.transport_lines = transport_lines
-        self.is_outskirts = is_outskirts
-        self.community_chest = community_chest
-        self.is_pier = is_pier
-        self.pier_type = pier_type
-        self.is_tutorial = is_tutorial
-        self.has_ads = has_ads
-        self.write_manuscript = write_manuscript
-        self.max_degradation = max_degradation
-        self.topic = topic
-        self.wikipage = wikipage
-
-        self.neighbors = neighbors
-        if self.neighbors == None:
-            self.neighbors = {}
 
     #  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54
 
@@ -1455,7 +1222,7 @@ async def look(cmd):
                 enemies_resp,
                 soul_resp,
                 ("\n\n{}".format(
-                    ewcmd.weather_txt(cmd.guild.id)
+                    ewutils.weather_txt(cmd.guild.id)
                 ) if cmd.guild != None else ""),
                 ad_formatting
             ) #+ get_random_prank_item(user_data, district_data) # SWILLDERMUK
@@ -1495,7 +1262,7 @@ async def survey(cmd):
                 slimeoids_resp,
                 enemies_resp,
                 ("\n\n{}".format(
-                    ewcmd.weather_txt(cmd.guild.id)
+                    ewutils.weather_txt(cmd.guild.id)
                 ) if cmd.guild != None else "")
             ) #+ get_random_prank_item(user_data, district_data) # SWILLDERMUK
         ))
@@ -1603,7 +1370,7 @@ async def scout(cmd):
 
         detailed_enemies_resp = "You pick up the scent of the following enemies:\n"
         for enemy in enemies_in_district:
-            enemy_data = EwEnemy(id_enemy=enemy)
+            enemy_data = ewhunting.EwEnemy(id_enemy=enemy)
             detailed_enemies_resp += "\n**{}**\n".format(enemy_data.display_name)
             num_enemies += 1
 
@@ -1780,7 +1547,7 @@ def get_enemies_look_resp(user_data, district_data):
         enemies_resp = ""
     # enemies_resp += "You don't find any enemies in this district."
     elif num_enemies == 1:
-        found_enemy_data = EwEnemy(id_enemy=enemies_in_district[0])
+        found_enemy_data = ewhunting.EwEnemy(id_enemy=enemies_in_district[0])
 
         if found_enemy_data.identifier != '':
             identifier_text = " {}".format(":regional_indicator_{}:".format(found_enemy_data.identifier.lower()))
@@ -1796,7 +1563,7 @@ def get_enemies_look_resp(user_data, district_data):
     else:
         enemies_resp += "You notice several enemies in this district, such as\n"
         for i in range(len(enemies_in_district) - 1):
-            found_enemy_data = EwEnemy(id_enemy=enemies_in_district[i])
+            found_enemy_data = ewhunting.EwEnemy(id_enemy=enemies_in_district[i])
 
             if found_enemy_data.identifier != '':
                 if not ewcfg.gvs_active:
@@ -1814,7 +1581,7 @@ def get_enemies_look_resp(user_data, district_data):
 
             enemies_resp += ("{} **{}" + identifier_text + "**\n").format(threat_emote, found_enemy_data.display_name)
 
-        final_enemy_data = EwEnemy(id_enemy=enemies_in_district[num_enemies - 1])
+        final_enemy_data = ewhunting.EwEnemy(id_enemy=enemies_in_district[num_enemies - 1])
 
         if final_enemy_data.identifier != '':
             if not ewcfg.gvs_active:
