+++ ew/prank.py
@@ -5,84 +5,6 @@ from . import utils as ewutils
 from .user import EwUser
 #from .player import EwPlayer
 
-"""
-	Prank items for Swilldermuk
-"""
-class EwPrankItem:
-	item_type = "item"
-	id_item = " "
-	
-	
-	alias = []
-	
-	context = "prankitem"
-	str_name = ""
-	str_desc = ""
-	
-	prank_type = "" # Type of prank item. Can be an instant use, trap, or response item
-	prank_desc = "" # A line of text that appears when the prank item gets used
-	rarity = "" # Rarity of prank item. Used in determining how often it should spawn
-	gambit = 0 # Gambit multiplier
-	response_command = "" # All response items need a different command to break out of them
-	response_desc_1 = "" # Response items contain additonal text which is indicative of how far the prank has progressed.
-	response_desc_2 = ""
-	response_desc_3 = ""
-	response_desc_4 = ""
-	trap_chance = 0 # All trap items only have a certain chance to activate
-	trap_stored_credence = 0 # Trap items store half your current credence up front for later
-	trap_user_id = "" # Trap items store your user id when you lay them down for later
-	side_effect = "" # Some prank items have side-effects. Example: The 'bungis beam' will change a player's name to '[player name] (Bungis)'
-	
-	ingredients = ""
-	acquisition = ""
-	vendors = []
-
-	def __init__(
-		self,
-		id_item=" ",
-		alias = [],
-		str_name = "",
-		str_desc = "",
-		prank_type = "",
-		prank_desc = "",
-		rarity = "",
-		gambit = 0,
-		response_command = "",
-		response_desc_1 = "",
-		response_desc_2 = "",
-		response_desc_3 = "",
-		response_desc_4 = "",
-		trap_chance = 0,
-		trap_stored_credence = 0,
-		trap_user_id = "",
-		side_effect = "",
-		ingredients = "",
-		acquisition = "",
-		vendors = [],
-	):
-		self.item_type = "item"
-		self.id_item = id_item
-		self.alias = alias
-		self.context = "prankitem"
-		self.str_name = str_name
-		self.str_desc = str_desc
-		self.prank_type = prank_type
-		self.prank_desc = prank_desc
-		self.rarity = rarity
-		self.gambit = gambit
-		self.response_command = response_command
-		self.response_desc_1 = response_desc_1
-		self.response_desc_2 = response_desc_2
-		self.response_desc_3 = response_desc_3
-		self.response_desc_4 = response_desc_4
-		self.trap_chance = trap_chance
-		self.trap_stored_credence = trap_stored_credence
-		self.trap_user_id = trap_user_id
-		self.side_effect = side_effect
-		self.ingredients = ingredients
-		self.acquisition = acquisition
-		self.vendors = vendors
-		
 # A bit of a hack, but at the time I couldn't really import cfg without causing other issues, so here these go. 
 # It's really the only place they get used anyways.
